02_main/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
PyJWT[crypto]==2.10.1
Pillow==10.4.0
python-hwpx
polar-sdk==0.30.1
standardwebhooks==1.0.1
